+++ lib/routemaster/cache_key.rb
@@ -1,7 +1,9 @@
 module Routemaster
   class CacheKey
+    PREFIX = 'cache:'.freeze
+
     def self.url_key(url)
-      "cache:#{url}"
+      "#{PREFIX}#{url}"
     end
   end
 end